refactor(services): move service classes into feature-specific packages; update package declarations accordingly
