+++ mark/src/main/java/pt/estga/mark/services/evidence/MarkEvidenceCommandService.java
@@ -1,4 +1,4 @@
-package pt.estga.mark.services;
+package pt.estga.mark.services.evidence;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
