+++ mark/src/main/java/pt/estga/mark/services/mark/MarkQueryService.java
@@ -1,21 +1,21 @@
-package pt.estga.mark.services;
+package pt.estga.mark.services.mark;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import pt.estga.mark.entities.MarkOccurrence;
-import pt.estga.mark.repositories.MarkOccurrenceRepository;
+import pt.estga.mark.entities.Mark;
+import pt.estga.mark.repositories.MarkRepository;
 
 import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class MarkOccurrenceQueryService {
+public class MarkQueryService {
 
-    private final MarkOccurrenceRepository repository;
+    private final MarkRepository repository;
 
-    public Optional<MarkOccurrence> findById(Long id) {
+    public Optional<Mark> findById(Long id) {
         return repository.findById(id);
     }
 }
